requirements.txt
# Context-Aware Robot Control System - Dependencies
# Python >= 3.10 required

# LangGraph Workflow
langgraph>=0.2.0
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-ollama>=0.1.0

# LLM Providers
openai>=1.0.0           # OpenAI API client
requests>=2.31.0        # HTTP client for Ollama API

# Knowledge Layer
neo4j>=5.14.0           # Graph database driver

# Configuration & Data Formats
python-dotenv>=1.0.0    # Environment configuration
pyyaml>=6.0.0           # YAML parsing (robot sequences)
